+++ crates/ewwii/src/widgets/build_widget.rs
@@ -41,6 +41,7 @@ fn build_gtk_widget_from_node(
 
     let gtk_widget = match root_node {
         WidgetNode::Box { props, children } => build_gtk_box(props, children, widget_reg)?.upcast(),
+        WidgetNode::FlowBox { props, children } => build_gtk_flowbox(props, children, widget_reg)?.upcast(),
         WidgetNode::EventBox { props, children } => {
             build_event_box(props, children, widget_reg)?.upcast()
         }