+++ crates/rhai_impl/src/dyn_id.rs
@@ -39,6 +39,10 @@ impl WidgetNode {
                 props: with_dyn_id(props.clone(), parent_path),
                 children: process_children(children, parent_path, "box"),
             },
+            WidgetNode::FlowBox { props, children } => WidgetNode::FlowBox {
+                props: with_dyn_id(props.clone(), parent_path),
+                children: process_children(children, parent_path, "flowbox"),
+            },
             WidgetNode::Expander { props, children } => WidgetNode::Expander {
                 props: with_dyn_id(props.clone(), parent_path),
                 children: process_children(children, parent_path, "expander"),